+++ src/lib/browser/client.js
@@ -1,0 +1,5 @@
+import { createPagesBrowserClient } from "@supabase/auth-helpers-nextjs";
+
+const supabaseBrowser = createPagesBrowserClient();
+
+export default supabaseBrowser;
