+++ src/lib/checkUser.js
@@ -1,0 +1,6 @@
+import { redirect } from "next/navigation";
+
+export default function checkUser(session) {
+  if (!session?.access_token) return redirect("/login");
+  return session.user;
+}
