+++ src/lib/server/session.js
@@ -1,0 +1,9 @@
+import supabaseServer from "./supabaseClientServer";
+import checkUser from "./checkUser";
+
+export async function getSessionServer() {
+  const {
+    data: { session },
+  } = await supabaseServer().auth.getSession();
+  return checkUser(session);
+}
